requirements.txt
Django==4.0.1
django-adminlte3
gunicorn==20.1.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt
asgiref
atomicwrites
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow
pluggy==1.0.0
py==1.11.0
PyJWT
pyparsing==3.0.6
pytest==7.1.2
psycopg2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe==8.5.0
toml==0.10.2
urllib3==1.26.7
